requirements.txt
aioredis==1.3.1
asgiref==3.3.4
astroid==2.8.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
boto3==1.18.53
botocore==1.21.53
certifi==2021.5.30
cffi==1.14.6
channels==3.0.0
channels-redis==3.2.0
chardet==4.0.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
defusedxml==0.7.1
Django==3.2
django-storages==1.11.1
djangorestframework==3.12.4
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isort==5.9.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
msgpack==1.0.2
oauthlib==3.1.1
Pillow==8.2.0
platformdirs==2.4.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pylint==2.11.1
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-decouple==3.4
python-social-auth==0.2.1
python3-openid==3.2.0
pytz==2021.1
razorpay==1.2.0
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
service-identity==21.1.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stripe==2.60.0
toml==0.10.2
twilio==6.59.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.5
wrapt==1.13.3
zope.interface==5.4.0
